Extract ci command from xtask main.rs
